chore: split oversized lint modules
