+++ packages/core/src/exports/platform.ts
@@ -1,0 +1,195 @@
+export { setBridge, hasBridge, type IpcBridge, type Unlisten } from '../ipc/bridge'
+export { call, callBinary } from '../ipc/invoke'
+export {
+  getAppVersion,
+  getAppPlatform,
+  isMobilePlatform,
+  mobileStorage,
+  mobileStorageLocal,
+  icloudDownloadPending,
+  icloudPendingCount,
+  icloudStatus,
+  icloudAdoptGraph,
+  icloudConflictsScan,
+  icloudWatchStart,
+  icloudWatchStop,
+  type AppPlatform,
+  type IcloudDownloadScope,
+  type IcloudScanOptions,
+  type IcloudStatus,
+  type IcloudSweepOutcome,
+  type MobileStorageInfo,
+  type MobileStorageKind,
+} from '../ipc/commands'
+export { confirmQuit, subscribeQuitRequested } from '../app/quit'
+export { toggleDevtools } from '../app/devtools'
+export {
+  agentSkillStatus,
+  agentSkillInstall,
+  agentSkillUninstall,
+  type AgentSkillInstallState,
+  type AgentSkillStatus,
+} from '../app/agent-skill'
+export { chunkNote, type NoteChunk } from '../embeddings/chunk'
+export {
+  embedStatus,
+  embedEnsure,
+  embedTexts,
+  embedApply,
+  embedRemove,
+  subscribeEmbedStatus,
+  embedStatusSchema,
+  type EmbedStatus,
+  type EmbedProgress,
+  type EmbedChunkPayload,
+} from '../embeddings/commands'
+export { embedNote, backfillEmbeddings } from '../embeddings/pipeline'
+export {
+  retrieve,
+  relatedNotes,
+  fuseRanked,
+  type RetrievalHit,
+  type RetrieveOptions,
+} from '../embeddings/retrieve'
+export {
+  appErrorSchema,
+  errorMessage,
+  isAppError,
+  toAppError,
+  ReflectError,
+  type AppError,
+} from '../errors'
+export {
+  DAILY_DIR,
+  NOTES_DIR,
+  TEMPLATES_DIR,
+  ASSETS_DIR,
+  AUDIO_MEMOS_DIR,
+  dailyPath,
+  notePath,
+  templatePath,
+  assetPath,
+  audioMemoPath,
+  descriptionPathFor,
+  DESCRIPTION_SUFFIX,
+  isDaily,
+  isNotePath,
+  isTemplatePath,
+  dateFromDailyPath,
+} from '../graph/paths'
+export {
+  graphInfoSchema,
+  recentGraphSchema,
+  fileMetaSchema,
+  graphImportSummarySchema,
+  type GraphInfo,
+  type RecentGraph,
+  type FileMeta,
+  type GraphImportSummary,
+} from '../graph/schemas'
+export {
+  openGraph,
+  createGraph,
+  importReflectV1Zip,
+  markReflectV1ImportOwnWrites,
+  readNote,
+  writeNote,
+  writeAsset,
+  readAsset,
+  openAsset,
+  listDir,
+  noteExists,
+  deleteNote,
+  listFiles,
+  recentGraphs,
+  forgetRecent,
+  deleteGraph,
+  captureHostRegister,
+  captureInboxList,
+  captureInboxRead,
+  captureInboxSpool,
+  captureInboxReject,
+  captureInboxRemove,
+  captureMetaFetch,
+  captureSharedInboxRelay,
+  promoteCaptureScreenshot,
+} from '../graph/commands'
+export { createAsset, importAsset } from '../graph/assets'
+export { assetFileName } from '../graph/asset-names'
+export {
+  newNoteId,
+  newNoteSource,
+  untitledNoteSeed,
+  untitledNotePath,
+  isUntitledNotePath,
+  createNoteWithTitle,
+} from '../graph/create-note'
+export {
+  settingsSchema,
+  editorMarkdownSyntaxSchema,
+  editorSpellCheckSchema,
+  editorDefaultBulletSchema,
+  editorBulletAfterHeadingSchema,
+  editorTextSizeSchema,
+  semanticSearchEnabledSchema,
+  describeAssetsSchema,
+  contactsEnabledSchema,
+  themePreferenceSchema,
+  timeFormatSchema,
+  dateFormatSchema,
+  weekStartDaySchema,
+  allNotesFilterTagsSchema,
+  calendarEnabledSchema,
+  calendarIdsSchema,
+  graphColorSchema,
+  graphColorsSchema,
+  GRAPH_COLOR_IDS,
+  aiProviderIdSchema,
+  aiProviderConfigSchema,
+  aiProvidersSchema,
+  defaultAiProviderIdSchema,
+  chatModelSelectionSchema,
+  aiPromptModeSchema,
+  aiPromptSchema,
+  aiPromptsSchema,
+  DEFAULT_SETTINGS,
+  type Settings,
+  type EditorMarkdownSyntax,
+  type EditorTextSize,
+  type ThemePreference,
+  type TimeFormat,
+  type DateFormat,
+  type WeekStartDay,
+  type AllNotesFilterTags,
+  type CalendarIds,
+  type GraphColor,
+  type GraphColors,
+  type AiProviderId,
+  type AiProviderConfig,
+  type AiPrompt,
+  type AiPromptMode,
+} from '../settings/schema'
+export { loadSettings, saveSettings } from '../settings/commands'
+export {
+  contactsAuthorizationSchema,
+  contactMatchSchema,
+  contactsAuthorizationStatus,
+  requestContactsAccess,
+  lookupContactsByEmail,
+  lookupContactsByName,
+  isContactsReadable,
+  type ContactsAuthorization,
+  type ContactMatch,
+} from '../contacts/commands'
+export {
+  contactLinkSuggestions,
+  contactNamesEqual,
+  matchContactForTitle,
+  suggestContactForTitle,
+} from '../contacts/match'
+export {
+  contactDetailsMarkdown,
+  appendContactDetails,
+  noteHasContactDetails,
+} from '../contacts/markdown'
+export { pickContactForEmail, resolveAttendeeContact } from '../contacts/resolve'
