+++ packages/core/src/exports/sync-markdown-indexing.ts
@@ -1,0 +1,214 @@
+export {
+  gitStatus,
+  gitSetup,
+  gitCommitAll,
+  gitFetch,
+  gitMergeRemote,
+  gitPush,
+  isDeviceFlowConfigured,
+  githubAppInstallUrl,
+  saveGithubAuth,
+  loadGithubAuth,
+  clearGithubAuth,
+  deviceFlowStart,
+  deviceFlowPoll,
+  runDeviceFlow,
+  getGithubToken,
+  gitDisconnect,
+  gitClone,
+  parseGithubRemote,
+  githubRemoteUrl,
+  newRepoUrl,
+  createGithubRepo,
+  getGithubRepo,
+  getAuthenticatedUser,
+  createGist,
+  updateGist,
+  deleteGist,
+  createSyncEngine,
+  isSyncError,
+  type GitStatus,
+  type CommitOutcome,
+  type MergeOutcome,
+  type PushOutcome,
+  type SkippedFile,
+  type ChangedFile,
+  type GithubAuth,
+  type DeviceFlowStart,
+  type DevicePollResult,
+  type GithubRepoRef,
+  type GithubRepo,
+  type GithubUser,
+  type PublishedGist,
+  type GistFile,
+  type SyncEngine,
+  type SyncEngineOptions,
+  type SyncErrorKind,
+  type SyncStatus,
+} from '../sync'
+export {
+  frontmatterSchema,
+  gistFrontmatterSchema,
+  gistBodyHash,
+  gistFilename,
+  isPinned,
+  pinnedOrder,
+  PARSED_NOTE_VERSION,
+  parseNote,
+  appendBlock,
+  appendUnderHeading,
+  appendTaskLine,
+  wikiLinkSafe,
+  editTaskLine,
+  removeTaskLine,
+  parseTaskMarker,
+  renameWikiLink,
+  setTaskDueDate,
+  clearTaskDueDate,
+  taskLineToBullet,
+  toggleTaskMarker,
+  TaskStaleError,
+  resolved,
+  resolveWikiLink,
+  resolveWikiLinkAsync,
+  unresolved,
+  splitFrontmatter,
+  parseFrontmatter,
+  upsertFrontmatter,
+  detectConflictMarkers,
+  resolveConflictMarkers,
+  conflictMarkerBlockCount,
+  conflictMarkerLabels,
+  type ConflictMarkerLabels,
+  parseBody,
+  reflectMarkdownParser,
+  wikiLinkExtension,
+  scanInlineWikiLinks,
+  scanInlineImages,
+  scanInlineSegments,
+  foldKey,
+  foldTag,
+  isTagName,
+  hasAuthoredTitle,
+  normalizeWikiTarget,
+  slugForTitle,
+  type ConflictResolution,
+  type Frontmatter,
+  type GistFrontmatter,
+  type Span,
+  type WikiLink,
+  type MarkdownLink,
+  type Heading,
+  type AssetRef,
+  type TaskMarker,
+  type ParsedNote,
+  type InlineWikiLink,
+  type InlineImage,
+  type InlineSegment,
+  type FrontmatterSplit,
+  type ParsedFrontmatter,
+  type NormalizedTarget,
+  type Resolution,
+  type WikiLookup,
+  type AsyncWikiLookup,
+} from '../markdown'
+export {
+  openIndex,
+  applyIndexedNote,
+  applyIndexedNotes,
+  removeFromIndex,
+  moveNoteIndexed,
+  clearIndex,
+  watchStart,
+  watchStop,
+  subscribeIndexChanges,
+  subscribeIndexApplied,
+  subscribeFileChanges,
+  emitFileChanges,
+  setLocalWriteEcho,
+  subscribeOwnWrites,
+  subscribeIcloudConflicts,
+  subscribeIcloudWatchFailed,
+  applyIndexChanges,
+  hashContent,
+  buildIndexedNote,
+  indexedNoteSchema,
+  indexedLinkSchema,
+  indexedAliasSchema,
+  indexNote,
+  reindexNotesReferencing,
+  rebuildIndex,
+  reconcileIndex,
+  syncIndex,
+  getIndexMeta,
+  setIndexMeta,
+  dailyDatesInRange,
+  getBacklinks,
+  getBacklinksWithContext,
+  getConflictedNotes,
+  getDuplicateNoteIds,
+  getLinkSources,
+  getNote,
+  getNotesByTag,
+  getOpenTasks,
+  getCompletedTasks,
+  groupTasks,
+  taskDateBucket,
+  listDailyNotes,
+  listNotes,
+  listNoteTags,
+  listRecentNotes,
+  getPinnedNotes,
+  searchNotes,
+  suggestWikiTargets,
+  suggestTags,
+  getIndexedFileFacts,
+  getIndexedFileFactsByPath,
+  resolveWikiTarget,
+  resolveNoteTarget,
+  rewriteLinksForTitleChange,
+  nextAliases,
+  availableNotePath,
+  availableTemplatePath,
+  slugPathForTitle,
+  templateSlugPathForTitle,
+  listTemplates,
+  type TemplateEntry,
+  parseHighlights,
+  randomNotePath,
+  parseSearchQuery,
+  searchWithFilters,
+  type IndexedNote,
+  type IndexedLink,
+  type IndexedAlias,
+  type Backlink,
+  type BacklinkContext,
+  type ConflictedNote,
+  type DailyNoteRow,
+  type DailyNotesRange,
+  type DuplicateIdGroup,
+  type NoteRow,
+  type OpenTask,
+  type TaskGroup,
+  type TaskGroupKind,
+  type NoteListEntry,
+  type NoteListOptions,
+  type NoteTagFacet,
+  type RecentNoteRow,
+  type RecentNotesOptions,
+  type PinnedNote,
+  type SearchHit,
+  type TagSuggestion,
+  type FileChange,
+  type WikiSuggestion,
+  type GeneratedDate,
+  type DateSuggestionContext,
+  type HighlightSegment,
+  type ParsedSearchQuery,
+  type SearchFilters,
+  type FilteredSearchHit,
+  type FilteredSearchOptions,
+  type RenameIo,
+  type TitleRenameRewriteOptions,
+  type TitleRenameRewriteResult,
+} from '../indexing'
